Fpminimax exhibitted an infinite loop in a particular case. When the coefficients are floating-point numbers, the algorithm has to guess the exponents of the coefficients first. Then, fpminimax is launched and we check afterwards that the coefficients have the same exponent as we guessed. If it is not the case, we correct the exponents and loop.
I knew that it could lead theoretically to an infinite loop. Yet, it was never oberved in practice.
Mioara found the following instance that effectively leads to an infinite loop:
f=sqrt(x); d=[1+12/512, 1+13/512];  fpminimax(f, 2, [|18, 10, 14|], d);

It turns out that the algorithm loops over the two polynomials
p=-1.00100001000101_2 * 2^(-3) * x^2 + 1.10010001_2  * 2^(-1) * x + 1.0110111010000111_2  * 2^(-2)
q=-1.100110000101_2  *  2^(-4) * x^2 + 1.011001011_2 * 2^(-1) * x + 1.10011011000101111_2 * 2^(-2)

Surprinsingly, the coefficients of x^2 is not the same at all in both polynomials.

I patched the algorithm with the following trick: when more than MAXLOOP trials have been performed, we stop as soon as a polynomial fits in the formats.
This still ensures that the result of the algorithm has the good form. This avoids the previous infinite loop but the theoretical possibility of an infinite loop stays.



This commit was extracted from SVN r786.

git-svn-id: svn://scm.gforge.inria.fr/svn/sollya/trunk@786 af37bc7e-59aa-4313-bbaf-7d19c0032f32