Corrected a bug in the addition/subtraction case in implementconstant. The bug possibly affected any addition/subtraction; however, it was particularly visible when the result of the addition was an exact zero.
Now, implementconstant complains (and aborts) whenever a subexpression is
exactly zero at a place where we cannot handle it.
WARNING: currently, this is very dirty, we raise an exception without freeing
anything.

Also added an example for libraryconstant (forgot from a previous commit).



This commit was extracted from SVN r1107.

git-svn-id: svn://scm.gforge.inria.fr/svn/sollya/branches/experimental-sylvain@1107 af37bc7e-59aa-4313-bbaf-7d19c0032f32