Corrected a bug in add/sub: there was subtle bug in my proof when the sufficient precision for the intermediate results was negative.
Probably, this had no influence on the code since the negative precision was replaced by MPFR_PREC_MIN anyway.
But it was safer to have a code corresponding to the proof. Moreover, the patch is an interesting optimization:
if the sufficient precision is negative, we can avoid computing the corresponding expression:
"(some complicated expr)+exp(100)" can be simplified to "exp(100)" up to a fairly big precision: there is no need
to evaluate the complicated expr in precision MPFR_PREC_MIN as we did before.

Interestingly, this led me to implement the if-then-else instruction: it is possible to write formal programs
containing branches. I also added a formal indentation mechanism for correctly displaying embedded blocks.

The code seems to be correct and no memory leak was observed. I did not perform intensive tests though.



This commit was extracted from SVN r1108.

git-svn-id: svn://scm.gforge.inria.fr/svn/sollya/branches/experimental-sylvain@1108 af37bc7e-59aa-4313-bbaf-7d19c0032f32