* Many tests were broken with recent versions of MPFR, due to the correction of the problem affecting mpfr_urandom (see commit message of revision 1270). As a consequence polynomials computed by remez algorithm were different with recent versions of MPFR compared to older versions and the tests were broken.
The set of tests should now pass both with old and recent MPFR libraries.

A problem might still remain with check012 (it is FPminimax). I have to investigate if it is 
just due to the remez thing or if there is something deeper.



This commit was extracted from SVN r1430.

git-svn-id: svn://scm.gforge.inria.fr/svn/sollya/trunk@1430 af37bc7e-59aa-4313-bbaf-7d19c0032f32