Removed a couple of evaluate*With/Without* function that were no longer in use.
Changed the behavior of evaluateFaithful, making quite a few tests fail.

The new evaluateFaithful returns the correct rounding to the nearest
as a faithful rounding wherever possible (if the proof interval is
sufficiently small) and returns the correct rounding to the nearest
(ties to even) of the midpoint of both possible answers otherwise. 

This change in behavior should remove any biais on the rounding direction
of faithful evaluations.

If this change in behavior does not satisfy the users (and test suite), it
may be reverted back in the future.




This commit was extracted from SVN r1709.

git-svn-id: svn://scm.gforge.inria.fr/svn/sollya/trunk@1709 af37bc7e-59aa-4313-bbaf-7d19c0032f32