* Updated sollya-messag*.h from branches/experimentalChebModels. For some cosmic reason, they had not been updated when I performed the svn merge at previous revision.
* Added check021.sollya.* and updated Makefile.am in directory tests-lib. Mioara used to
put them in the old 'tests' directory that does not exist anymore. As a consequence : the
tests do not pass currently, because of check021.sollya. I suspect that Mioara always thought
that her tests passed because her 'make check' was indeed not using her files.



This commit was extracted from SVN r1747.

git-svn-id: svn://scm.gforge.inria.fr/svn/sollya/trunk@1747 af37bc7e-59aa-4313-bbaf-7d19c0032f32