* Second attempt. The problem that we try to fix here is that adding -lstdc++ in the compilation line does not work when gcc is used to link and when gcc version does not agreee with g++ version. Also, it might not work if the version of g++ does not agree with the version of lstdc++.
We add -lstdc++ to the compilation line only if we are sure that it works. Not adding it does no harm in general, as it is only needed on Cygwin, with old versions of fplll.
Hopefully (but it is hard to test it in order to be certain), if we are in the unlikely case when:
  1) we have an old version of fplll
  2) we are on Cygwin
  3) we cannot add -lstdc++ to the compilation line,
configure will detect that sollya cannot be build on this specific machine. What makes me optimistic about it is the fact that there is a test after all this mess that tries to compile and run a small example program using fplll. I think that that test will fail on this particular architecture.




This commit was extracted from SVN r2200.

git-svn-id: svn://scm.gforge.inria.fr/svn/sollya/trunk@2200 af37bc7e-59aa-4313-bbaf-7d19c0032f32