First step towards uniform implementation of unary base functions.
After some review of the existing code, it seems more reasonable to consider NEG as an arithmetic operator than to consider it as unary function, because there is many cases where NEG is handled differently than the other unary functions.
Reviewing the code let me observe that the way the neg operator is displayed seems unconsistent accross several functions (sRawPrintTree, printTree, fprintf, etc.) I put two TODOs in execute.c and expression.c to remember that there is something to look at.

At many places, the code could be further reduced by collecting together several match cases. However, in order to separate tasks, I did not perform any modification that is not directly related to the new UNARY_BASE_FUNC feature.

It might be that the name 'base_func' is not ideally chosen as it conflicts with a similar name used in the name space of the library. However, this will be easy to change later, so I prefer to postpone this point to a further discussion.