Now comes the tricky part...
I exported the code used to evaluate the sign of a constant function
from infnorm.c into the new base-function.c. I simply copied the code
verbatim, but detected a possible inconsistency in the way different
positive functions are handled: in the case of exp(c), it seems that
there is a test that c evaluates to some true number, while for other
functions like cosh(c), no such test is performed.

Every non trivial operation that used to be performed in expression.c
is now reported on a specific function in the type baseFunction (e.g.,
differentiating an expression tree when the head is a base function,
simplifying expressions, etc.).

Regarding the point evaluation of a function, in most cases it
suffices to consider the corresponding mpfr function but for some
obscure reason, the functions double/single/quad/etc. are handled a
different way with some magic choice of the precisions (see for
instance this used to be the code that used to be in the function
evaluateConstantExpression). I am not understanding this fact very
well, and I decided to write special point_eval functions for these
cases, but I am not convinced that it is absolutely necessary. In any
case, as a result, the semantics of point_eval functions in a
basefunction structure is still not precise, so a choice shall still
be made and documented. I left some "TODO" marks here and there that
are related to this issue (and others as well. Generally speaking, it
will be necessary to grep TODO on the whole code before reporting this
branch anywhere).

Regarding the simplification of expressions, I factorized what was
obvious to factorize for functions having no special property. For the
others, I copied the code verbatim. However, I noticed some
inconsistencies in the way they are coded: for instance CEIL and FLOOR
have a subtle difference in the choice of precisions for some
variable, and it seems to me as an error of copy-paste. I am pretty
sure that the code for double/single/quad/etc. could be factorized as
well but there also, there are subtle choice of precisions that I am
not sure to understand, so I preferred touch nothing for now. Also, I
remarked that abs(c) could be simplified as soon as the sign of c
could be determined, but this is not performed.

Pretty much the same remarks apply to differentiation of functions:
functions that should reasonnably be differentiated in a similar
manner are indeed not. Some differentiations involve obscure tricks
(coding pi as 4*atan(1) for instance) that seem to come from ages
where sollya was a baby toy. Nevertheless, I simply copied the code at
it was in order to separate tasks and keep compatibility with the
existing tests. I just updated the code so as to make it more readable
by using makeFUNCTION functions instead of building the expressions
trees by hand.

Alltogether, these codes should definitely be reviewed, and the good
news is that everything is now concentrated in basefunction.c so that
it is easy to review it and compare the ways several comparable
functions are handled.

The code is now in a state where it can almost be compiled. It remains
just two occurrences of the old-style base functions in expression.c
and one occurrence in infnorm.c.