Some uncommitted optimization in Newton's algorithm that were forgotten.
We set up these optimization with Mark and Christoph a while ago but that I
forgot until now. If I remember well, these optimization did not change much
the efficiency of Newton's algorithm by that time, but it could be different
once the branch improvedEvaluate will be ported back onto the trunk.

Also removed an old wrapper that is now useless.