Last modifications of the code to centralize the definition of base functions.
Added boolean variables in the definition of base functions, in order to allow
their use in the remaining places of expression.c and infnorm.c. Still some
weird things discovered in the existing code (e.g., cosh considered as having a
zero at zero). I reproduced the existing behavior even though I hardly believe
it was intentional. See TODO mentions in the code.

Also, many many small debug things that appeared when I tried to compile.

The code is now in a state where it compiles with a few warnings about the mpfi
type functions. This is something we have to discuss: mpfi standard requires to
return an integer (that we never use) and our own implementation of mpfi-like
functions do not conform this standard. This is a minor issue and the warning
can be removed by simply deciding that interval_eval method returns void and
explicit cast to this type for those functions that indeed return an integer.

Everything seems to run smoothely. Only one check fails and I guess it could
come from one of the places where I left some TODOs to discuss. Also, two tests
fail in tests-lib, but I wonder if it is not already the case in master (I
still need to check this point).

Next steps are: do further tests, discuss all the TODOs, simplify and
optimize the old code that I simply copied while it could be optimized and
implement the derivatives of nearestint and co.