Patched one of the error reported by make check. The other one is under control.
It appears that it was only a stupid error of retranscription in a copy/paste.

Regarding the second error (check025), it comes from the fact that at one place
(and only one place) mpfr_rint_ceil was used to evaluate, although mpfr_ceil is
used everywhere else. The new code does not make this distinction (and I am not
sure that there is any reason to do so. This is something we should discuss)
and this leads to a subtle change in the behavior of sollya ceil function when
the precision does not allow to exactly represent the result of the
computation. Indeed, mpfr_rint_ceil acts as for any other mathematical
function: so, if used with GMP_RNDN, it returns the representable integer that
is the closest to the true value of ceil(x). The unpleasant consequence being
that this integer can be smaller than x, which is counter-intuitive. In
contrast, mpfr_ceil returns the smallest representable integer that is greater
or equal to the true value of ceil(x) which can be less accurate, but is more
intuitive. The problem is in the semantics of the returned value of mpfr_ceil:
if I understand well the documentation, and if x is not an integer, the
returned value does not tell anything about the exactness of the computed
value. It only says that x is not a mathematical integer. And the thing that we
care about for simplifying an expression of the form ceil(a) is precisely to
know if ceil(a) has been exactly computed or not, and not if a itself is an
integer. A possibility could be to use mpfr_rint_ceil with GMP_RNDU, which
should behaves excatly as mpfr_ceil does, except for the returned value. So,
IMHO we should disuss the question of keeping a rnd argument in our eval API
for base-functions and use mpfr_rint_ceil in any case.