Corrected kind of a bug in configure when using --with-package without argument
When --with-package is given to configure without any "=something" argument, the value of the argument defaults to "yes" which is non-sense in the purpose with which we use the --with-package option when configuring Sollya.

Moreover, the extra "-Lyes" produced by configure in this case turns out to be a problem at compile time (though not at configure time) because libtool seems to be extra careful about it and checks that the directory given as argument of "-L" indeed exists.

This commit patch this behavior by simply ignoring the --with-package directive when no argument or an empty argument is given.