Attempt to remove a warning when compiling under Ubuntu.
Namely, on Ubuntu, I got the following warning:
  ar: `u' modifier ignored since `D' is the default (see `U')

Accroding to the discussion on the following thread
http://gnu-automake.7480.n7.nabble.com/bug-20082-new-warning-from-ar-on-rawhide-systems-td21769.html
on the mailing list of automake, a safe patch seems to simply and manually set ar flags to 'cr' which I believe I do with this patch.

Compiles without a warning on Ubuntu and my Fedora.