Corrected a bug in shlp2* scripts
When simulating an example whose code contains escaped characters such as '\n' or '\"', the string was evaluated which was of course not intended. As a result, in some case, the output of the simulation was wrong.

This should now be fixed. After checking, the only file related to the bug was objectname.shlp (which was the file that made me discover the bug).