Added support for skipping lines of output in the examples of shlp files.
In a #EXAMPLE section of a shlp file it is now possible to add a comment
of the form /* Skip NN */ at the end of a line. The first NN lines of the
output produced by this specific line will be removed and a warning will be
displayed explaining that some lines have been removed.

This feature is quite fragile because of the constraints imposed by the
Verbatim environment in LaTeX. It will most probably fail if the (non removed
part of the) output contains one of the characters '\', '~' or '|'.