Reported spelling corrections into doc source files.
Based on the suggestions made by Jerome Benoit for the Debian team on previous commits.

We accept all proposed corrections except for ChangeLog which is an automatically generated file from the commit messages: if there are spelling errors in the commit messages, that is too bad, but it is the way it is and there is no reason to change it in the generated ChangeLog.