Removed lines that have been commented by Jerome Benoit.
If they are useless, there is no need to keep them as a comment.
If they turn out to be useful for some reason, we can always revert the change.

Also, I did not remove the carriage return introduce by Jerome Benoit in Makefile.am before the -version-info flag. I assume he introduced it for some reason (maybe to prepare the use of version script?). I wonder why he removed $(AM_LDFLAGS) however. I would think we introduced it for some reason in the first place. The proof of the pudding is in the eating: let us see if we manage to compile everywhere without this parameter.