Fixed a bug introduced with 8d2d75bbafda71eb5446fcc9b0f96383a47ab370
In that commit, Jerome Benoit removed AM_LDFLAGS from the definition of libsollya_la_LDFLAGS. I do not know if this was intentional or accidental, but it turns out that this was necessary, e.g. to link against a library installed in a non-conventional place specified through the --with-xxx argument of configure.