Less fragile support for timecheck
timecheck is designed to be quite robust, but, by its nature may fail on some unusual architectures. This could be misleading to people that might think that they have a problem with their compilation of Sollya.

An idea proposed by Jerome Benoit was to list it as XFAIL, but timecheck *is not* expected to fail: it is expected to pass, although, sometimes, it might not pass (notice that these situations are worth being reported). As a compromise, I propose in this patch simply to mark it as skipped when it does not pass.