Removed libsollyaexample.c
This file had been created initially to offer a panorama of how to used the library but it has never been really filled with something non-trivial.

The idea remains good in itself but would require some work. Currently, the official documentation does the job and the test-lib/*.c files offer a lot of practical examples, more valuable than the current libsollyaexample.c