Time-complexity checks in tests-lib independent from Sollya library
In order to avoid declaring symbols in sollya.map that are useless except for the checks of the lib, I propose the following patch:
  - separate tests of tget_element_in_list into two separate files: one file is specifically for checking the time complexity.
  - this time check is performed only on systems where gettimeofday is available

Right now, if gettimeofday is available the test is performed and may fail (as opposed to tests-tool/timecheck.sollya that is marked as 'SKIP' instead of 'FAIL' when it does not pass).
Normally, the check for tget_element_in_list_timecomplexity is designed to be really robust and measure a theoretical complexity, so I am in favor of really marking it as 'FAIL' in case it does not pass. But notice that this behavior is easy to change if we wish.

Normally, I should have removed everything that had been added to the source code specifically to make sollya_gettime available and usable in the tests of the lib.