(Kind of) fixing bug [#8867] Possible infinite loop in Remez
This led me to introduce a new argument to uncertifiedInfnorm that allows one to store an argmax.

The patch is a patch over a patched algorithm, and all this does not really make sense any more. At least, this patch makes both examples reported in the bug report terminate. I have no proof of any kind that it completely patches the problem.