Attempt to patch configure issue with fplll on Cygwin.
The changes introduced at commit ea4ede2 break configure with some versions of fplll on Cygwin. Indeed, the option -std=g++11 was now systematically added when supported by the compiler, but it turns out that the compilation of some older versions of fplll under Cygwin raise an error when compiled with this option, although they compile fine without the option...

This commit attempts to fix this issue by adding the -std=g++11 option only when the new API is detected, indicating that fplll is recent enough.