Removed code that seems useless.
While investigating the bug addressed by Christoph at commit e0b3184, I found a piece of code I could not understand the logic of in infnorm.c.
This code adds a complementary test to an existing test deciding whether we should run Newton's algorithm within dirtyinfnorm algorithm. The addition has been made by Christoph at commit 9b1d5e6. This commit performs many non trivial modifications on many files, including infnorm.c, but this precise modification is completely isolated (in particular, it is the only modification to the dirtyinfnorm algorithm). There is no comment that would provide us with an idea of the intent behing this modification, and the commit log is not more helpful.

I cannot see any logic in this modification, although it is rather harmless in the sense that it would simply conclude that it is necessary to run Newton's algorithm in situations where it does not seem quite useful.

The present commit reverts this modification. All the tests keep passing.