Removed definitions of sollya_C*FLAGS in Makefile.am (they seem useless).
Christoph introduced them when we started compiling the interactive tool from the library. He could not come with any precise argument why he put these definitions, except that he remembers having copied it from some more or less official example found on the web.

As far as I understand the logic of automake, this is useless because AM_C*FLAGS are automatically added to any target. I tried removing it and compiling sollya on a computer where the usual libs are installed in some prefix, providing them through the --with- mechanism and it worked perfectly.