Kind of patched bug #12892.
The patch is not perfect, but works and will work in most situations, although this is not the smartest thing we could do. The problem was that multi-point exchange algorithm sometimes behaves badly when Haar condition is not fulfilled, leading to a situation where epsilon could suddenly decrease (instead of constantly increase), therefore leading to possible pseudo-cycles in the algorithm.

When we detect that epsilon has decreased, we now simply switch to the single point exchange algorithm and keep this choice active until the end of the algorithm. This ensures that no cycle can happen.

We could do better by starting over from the previous set of points (instead of simply continuing with the set of points that led to epsilon decreasing) but the patch to do it would be more involved, and in practice it does not seem to be very harmful to simply continue the algorithm with the points obtained after epsilon has decreased. Indeed, this patch also fixes bug #17200 except for one case where, precisely, it is not a good idea to continue the algorithm after epsilon has decreased.