Completely reworked the configure tests for FPLLL.
The recent change of API in fplll introduce new combinatoric possibilities: one might or now want to add the -std=c++11 option, and one might or not use the new API. This is to add to the old possibilities regarding the position of fplll.h in the include directory tree and the question of whether -lfplll flag should be passed or not the linker.

At first, I thought that things were not much more difficult: the new -std requirement comes with the new API and that's it. It turns out that fplll API is currently not stable and the new API for instance was the one to use with fpllll 5.0 but not with the current git repository of fpllll. Nevertheless, the requirement of -std=c++11 is stays present with the current repository, although the old API is used.

To make things more complicated, we have examples of versions of fplll (observed when trying to compile under Christoph's Cygwin system) where passing -std=c+11 breaks the compilation while not passing it works fine. Also, recently Victor Magron signaled that versions of g++ around 4.8 and 4.9 works fine with modern versions of fplll *only* is the option -std=c++11 is explicitely passed (although with more recent versions of g++, it compiles fine with or without the option).

So, I decided to more or less systematically test the possible combinations of these four parameters. Some combinations really seem to be unlikely because they refer to behaviors of fplll at very different time of its history. So we indeed try only 7 possibilities over the 16 possibilities. Hopefully, this should allow us to compile sollya with more or less any version of fplll.