Replaced calls to rand_r by calls to rand (rand_r is marked as obsolete from POSIX.1-2008).
I am not sure that this is related to the fact that it is obsolete, but we had a warning 'implicitely declared function' under Debian about rand_r which was mysterious. It did not appear clearly to me that rand_r was really necessary in our context, so I switched to the couple srand/rand.